+++ src/main/java/org/example/proxishop/model/entities/customer/Customer.java
@@ -9,14 +9,14 @@ import lombok.Setter;
 public class Customer {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Double id;
+    private int id;
     private String lastName;
     private String firstName;
     private String email;
-    private Double id_shoppingCart;
+    private int id_shoppingCart;
 
 
-    public Customer(Double id, String lastName, String firstName, String email, Double id_shoppingCart) {
+    public Customer(int id, String lastName, String firstName, String email, int id_shoppingCart) {
         this.id = id;
         this.lastName = lastName;
         this.firstName = firstName;
@@ -24,7 +24,7 @@ public class Customer {
         this.id_shoppingCart = id_shoppingCart;
     }
 
-    public Customer(String lastName, String firstName, String email, Double id_shoppingCart) {
+    public Customer(String lastName, String firstName, String email, int id_shoppingCart) {
         this.lastName = lastName;
         this.firstName = firstName;
         this.email = email;
